+++ goodwatch-webapp/app/server/dna.server.ts
@@ -25,8 +25,8 @@ export const getDNA = async (params: DNAParams) => {
 		name: "dna",
 		target: _getDNA,
 		params,
-		// ttlMinutes: 60 * 24,
-		ttlMinutes: 0,
+		ttlMinutes: 60 * 24,
+		// ttlMinutes: 0,
 	})
 }